remove eclipse config files
The eclipse project and config files are very workplace specific and often cause trouble and merge conflicts when shared with contributors. Therefore it is easier to share the project as a maven project which can be imported easily by m2e.

This makes it also unnecessary to ship the libraries in the lib directory with the repository, as m2e will automatically download the dependencies from the maven repository. As however this is not supported by ant, I won't remove the lib directory. Removing ant support should be considered.