add withCopyMethod to GeneratePojoBuilder
When withCopyMethod is true, a special method named copy will be generated.

This method allows you to easily create a new object that is identical to another object.

That is especially helpful, when you want to create a new object from an already existing immutable object, that has only a few fields that differ.